backend/requirements.txt
django>=5.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.2
python-dateutil>=2.9.0.post0
